implemented ResultsQueryer

+++ src/main/java/com/jwm/ir/search/Document.java
@@ -1,7 +1,7 @@
 package com.jwm.ir.search;
 
 public interface Document {
-	int getDocumentId();
+	long getDocumentId();
 	int getLength();
 	int getTermFrequency(String term);
 }
